q2_prompt_engineering_pipeline/requirements.txt
google-generativeai
python-dotenv
pydantic
numpy
pandas
matplotlib
seaborn
scikit-learn
tqdm
loguru
tenacity